Why autosetting frame is dumb
If the frame is autoset with the moving pixels, the moment there is no drop it will be set to 0. Although that's techincall not true since every time the frame changes everything moves. This means that it will slowly shrink to nothing. The solution? Corner averaging? Boudning volume hierarchy? Template sliding?